+++ database/migrations/2024_05_09_191912_create_localidades_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('cat_localidades', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('municipio_id')->unsigned()->index();
+            $table->mediumInteger('municipio_id')->unsigned()->index();
             $table->string('nombre', 100);
             $table->string('nombre_corto', 50);
             $table->string('codigo', 10);
